+++ outputs.tf
@@ -1,13 +1,3 @@
-output "cluster_username" {
-  description = "Cluster master username, keep always secret!"
-  value       = random_string.cluster_username.result
-}
-
-output "cluster_password" {
-  description = "Cluster master password, keep always secret!"
-  value       = random_string.cluster_password.result
-}
-
 output "endpoint" {
   description = "Cluster control plane endpoint"
   value       = google_container_cluster.primary.endpoint
@@ -20,21 +10,22 @@ output "node_pools" {
 
 output "client_certificate" {
   description = "Client certificate used kubeconfig"
-  value = base64encode(
-    google_container_cluster.primary.master_auth[0].client_certificate,
-  )
+  value       = base64decode(google_container_cluster.primary.master_auth[0].client_certificate)
 }
 
 output "client_key" {
   description = "Client key used kubeconfig"
-  value       = base64encode(google_container_cluster.primary.master_auth[0].client_key)
+  value       = base64decode(google_container_cluster.primary.master_auth[0].client_key)
 }
 
 output "cluster_ca_certificate" {
   description = "Client ca used kubeconfig"
-  value = base64decode(
-    google_container_cluster.primary.master_auth[0].cluster_ca_certificate,
-  )
+  value       = base64decode(google_container_cluster.primary.master_auth[0].cluster_ca_certificate)
+}
+
+output "access_token" {
+  description = "Client access token used kubeconfig"
+  value       = data.google_client_config.default.access_token
 }
 
 output "instance_group_urls" {
@@ -46,4 +37,3 @@ output "cluster_ipv4_cidr" {
   description = "The IP address range of the Kubernetes pods in this cluster in CIDR notation"
   value       = google_container_cluster.primary.cluster_ipv4_cidr
 }
-